+++ src/app/(routes)/(tabs)/profile/edit/page.tsx
@@ -1,7 +1,7 @@
 import TitleHeader from '@/shared/components/title-header'
-import EditForm from './components/edit-form'
+import EditForm from '@/views/profile-edit'
 
-export default function ProfileEdit() {
+const Page = () => {
   return (
     <main className="h-[calc(100vh-84px)] bg-white lg:h-screen">
       <TitleHeader title="프로필 정보 수정" />
@@ -11,3 +11,5 @@ export default function ProfileEdit() {
     </main>
   )
 }
+
+export default Page
