+++ src/app/(routes)/(tabs)/profile/notification/email/page.tsx
@@ -1,7 +1,7 @@
 import TitleHeader from '@/shared/components/title-header'
-import EmailSettingForm from './components/email-setting-form'
+import EmailSettingForm from '@/views/notification-email'
 
-export default function ProfileEmail() {
+const Page = () => {
   return (
     <main className="h-[calc(100vh-84px)] bg-white lg:h-screen">
       <TitleHeader title="퀴즈 알림 이메일 설정" />
@@ -11,3 +11,5 @@ export default function ProfileEmail() {
     </main>
   )
 }
+
+export default Page
