+++ src/views/profile/config.ts
